#649 - First pass on the refactoring of the webserver
